The great re-write begins...
A lot of dates will be marked incorrectly, mainly because I can't be bothered to open each file and manually update the date, and also because I'll have to update each file anyway to account for the new styles.

Still need to update the templates to account for some of the new things.